+++ tests/test013_calvin_processor.py
@@ -70,3 +70,69 @@ def test_mwax_calvin_processor():
         == "--uvw-min 50m --uvw-max 1667l --max-iterations 500 --stop-thresh 1e-20 --veto-threshold 0.0 --source-dist-cutoff 50"
     )
     assert mcal.phase_fit_niter == 3
+
+
+def test_mwax_calvin_processor_no_gains_cutoff():
+    """Tests that mwax_calvin reads a config file ok"""
+    # Setup all the paths
+    base_dir = tests_common.setup_test_directories("test013")
+
+    # Start mwax_subfile_distributor using our test config
+    mcal = MWAXCalvinProcessor()
+    mcal.produce_debug_plots = False
+
+    # Override the hostname
+    mcal.hostname = "calvin99"
+
+    # Determine config file location
+    config_filename = "tests/data/test013/test013_no_gains_cutoff.cfg"
+
+    # Override db_handler with a fake one
+    fake_db_handler = FakeMWAXDBHandler()
+    # Add any select results (in order in the code below-or keep commented if none)
+    # e.g. fake_db_handler.select_results = [[{"observation_num": 123, "size": 1024, "checksum": "abc123"}]]
+
+    # Call to read config <-- this is what we're testing!
+    mcal.initialise(
+        config_filename,
+        1234567890,
+        1234,
+        None,
+        CalvinJobType.realtime,
+        "https//test.test",
+        [
+            123,
+        ],
+        fake_db_handler,
+    )
+
+    #
+    # Now confirm the params all match the config file
+    #
+
+    # mwax_mover section
+    assert mcal.log_path == os.path.join(base_dir, "logs"), (
+        f"log path mismatch: {mcal.log_path} {os.path.join(base_dir, 'logs')}"
+    )
+
+    assert mcal.health_multicast_interface_name == "eth0"
+    assert mcal.health_multicast_ip == "127.0.0.1"
+    assert mcal.health_multicast_port == 8011
+    assert mcal.health_multicast_hops == 1
+
+    assert mcal.job_input_path == os.path.join(base_dir, "data/calvin/in_jobs")
+    assert mcal.job_output_path == os.path.join(base_dir, "data/calvin/out_jobs")
+    assert mcal.source_list_filename == "../srclists/srclist_pumav3_EoR0aegean_fixedEoR1pietro+ForA_phase1+2.txt"
+    assert mcal.source_list_type == "rts"
+    assert mcal.hyperdrive_binary_path == "../mwa_hyperdrive/target/release/hyperdrive"
+    assert mcal.hyperdrive_timeout == 7200
+    assert mcal.birli_binary_path == "../Birli/target/release/birli"
+    assert mcal.birli_timeout == 3600
+    assert mcal.keep_completed_visibility_files == 0
+    assert mcal.cal_export_max_age_hours == 24
+    assert (
+        mcal.hyperdrive_extra_args
+        == "--uvw-min 50m --uvw-max 1667l --max-iterations 500 --stop-thresh 1e-20 --veto-threshold 0.0 --source-dist-cutoff 50"
+    )
+    assert mcal.phase_fit_niter == 3
+    assert mcal.gains_cut_off_max == -1.0