+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+# 1.7.8 19-May-2026
+
+* calvin: Added missing "Preparing" ASVO state to list of valid states.
+* calvin_processor: Disable gains cut off if value is not set or negative.
+
 # 1.7.7 19-May-2026
 
 * calvin_processor: Now sets gains of entire jones matrix to NaN if gain of any part of the matrix is > cut off specified in config file.